Step 2
Populate corpus directory with rolled-up sample files for case
statements or themata, dicta from the marriage causae (C.27-C.36),
and dicta from de Pen., all from the first recension. Because the
only goal at this point is to generate a starter wordlist, the
contents of the sample files aren't in order. These samples have
not been processed to detach the -que enclitic.